When you edit the length of a time slot after it's first created, the times in the Visit sign-up header are incorrect
handleVisitChange was converting a height in pixels into a date, which doesn’t work. That’s why usually the date shown after changing a visit was early in the morning because heights are usually short (i.e. lower numbers), and mornings are lower numbers too. This fix makes passes the bottom pixel dimension, instead of the height, and calculates the height in handleVisitChange where needed