If you drag a time slot off the bottom of the calendar it disappears and an error appears in the logs
It turns out in handleVisitChange we didn't handle the case of the top or bottom being exactly equal to calendarTopTime or calendarBottomTime. Thus updateVisit() was called with undefined for the start and end times, and it would disappear from the data set for some reason.

This fix also fires a warning if we can't determine the start and end time and returns early without modifying the record